requirements.txt
# ── Core dependencies ─────────────────────────────────────────────────────────
# Config parsing
pyyaml>=6.0

# HTTP client (OpenAI-compatible API backend)
httpx>=0.27.0

# Numerical utilities (used in analysis/plotting)
numpy>=1.26.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0

# ── Local model serving (optional) ───────────────────────────────────────────
# Required only when running locally-served models via TransformersQwenBackend.
# API-accessed models (DeepSeek, GLM-5, Qwen3.6-plus, Minimax) do NOT need these.
# Uncomment if you plan to serve models locally with llama.cpp or transformers:
#
# transformers>=4.40.0
# torch>=2.2.0
# accelerate>=0.29.0

# ── External planner ─────────────────────────────────────────────────────────
# Fast Downward is NOT installed via pip. Install separately:
#   git clone https://github.com/aibasel/downward.git ~/downward
#   cd ~/downward && python build.py
# Then invoke as: ~/downward/fast-downward.py <domain> <problem> --search "astar(blind())"
